backend/requirements.txt
django
djangorestframework
python-dateutil
Pillow

./external/InvoiceGenerator
